pyproject.toml
[project]
name = "bdpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.50.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
]
